requirements.txt
praw
tqdm
dotenv
datetime
ansicolors